git-svn-id: svn://svn.r-forge.r-project.org/svnroot/gwatoolbox@53 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/src/gwafile/GwaFile.cpp
@@ -81,7 +81,7 @@ char GwaFile::get_data_separator() {
 	return data_separator;
 }
 
-unsigned int GwaFile::get_estimated_size() {
+unsigned long int GwaFile::get_estimated_size() {
 	return estimated_size;
 }
