+++ pkg/src/gwasqc/include/metas/MetaCrossTable.h
@@ -58,7 +58,7 @@ public:
 	bool is_na();
 	void print(ostream& stream);
 	void print_html(ostream& stream, char path_separator);
-	int get_memory_usage();
+	double get_memory_usage();
 };
 
 #endif
